fix(fe|layout): 修正前端布局
- 添加dashboard和user两个虚拟路由，避免登录页出现侧边栏
