+++ frontend/app/(dashboard)/instrument/components/ActionButtons.js
