+++ frontend/app/(dashboard)/instrument/components/Instrument_Card.js
