+++ frontend/app/(dashboard)/instrument/components/Status.js
