+++ frontend/app/(dashboard)/instrument/page.js
